added quiz generation backend logic
