+++ client/src/app/app-store/functions/index.ts
@@ -1,0 +1,9 @@
+export * from "./generate-quiz";
+export * from "./shuffle-array";
+export * from "./determine-quiz-display";
+export * from "./are-key-words-present";
+export * from "./grade-open-ended-answer";
+export * from "./grade-specific-question";
+export * from "./is-answer-similar";
+export * from "./levenshtein-distance";
+export * from "./shuffle-array";
