+++ client/src/app/app-store/index.ts
@@ -1,0 +1,2 @@
+export * from "./functions";
+export * from "./mocks";
