--- client/src/app/components/MockTrueFalseQuestions.tsx
@@ -1,79 +1,0 @@
-
-
-export const mockTrueFalseQuestions = [
-  {
-      question: 'The Earth is flat.',
-      options: ['true', 'false'],
-      answer: "false"
-  },
-  {
-      question: 'Water boils at 100°C.',
-      options: ['true', 'false'],
-      answer: "true"
-  },
-  {
-      question: 'The sun revolves around the Earth.',
-      options: ['true', 'false'],
-      answer: "false"
-  },
-  {
-      question: 'Humans need oxygen to survive.',
-      options: ['true', 'false'],
-      answer: "true"
-  },
-  {
-      question: 'Sharks are mammals.',
-      options: ['true', 'false'],
-      answer: "false"
-  },
-  {
-      question: 'Lightning never strikes the same place twice.',
-      options: ['true', 'false'],
-      answer: "false"
-  },
-  {
-      question: 'Gold is a metal.',
-      options: ['true', 'false'],
-      answer: "true"
-  },
-  {
-      question: 'The Pacific Ocean is larger than the Atlantic Ocean.',
-      options: ['true', 'false'],
-      answer: "true"
-  },
-  {
-      question: 'Bats are blind.',
-      options: ['true', 'false'],
-      answer: "false"
-  },
-  {
-      question: 'Electrons are larger than atoms.',
-      options: ['true', 'false'],
-      answer: "false"
-  },
-  {
-      question: 'Penguins can fly.',
-      options: ['true', 'false'],
-      answer: "false"
-  },
-  {
-      question: 'Cats are reptiles.',
-      options: ['true', 'false'],
-      answer: "false"
-  },
-  {
-      question: 'The Great Wall of China is visible from space.',
-      options: ['true', 'false'],
-      answer: "false"
-  },
-  {
-      question: 'The heart is a muscle.',
-      options: ['true', 'false'],
-      answer: "true"
-  },
-  {
-      question: 'Venus is the hottest planet in the solar system.',
-      options: ['true', 'false'],
-      answer: "true"
-  },
-];
