+++ client/src/app/components/index.tsx
@@ -1,0 +1,12 @@
+export { default as CheckButton } from "./CheckButton";
+export { default as CheckQuizHistoryButton } from "./CheckQuizHistoryButton";
+export { default as DownloadQuiz } from  "./DownloadQuiz";
+export { default as GenerateButton } from "./GenerateButton";
+export { default as NewQuizButton } from "./NewQuizButton";
+export { default as QuizAnswerField } from "./QuizAnswerField";
+export { default as QuizForm }from "./QuizForm";
+export { default as QuizGenerationSection } from "./QuizGenerationSection";
+export { default as SignInButton } from "./SignInButton";
+export { default as SignInModal } from "./SignInModal";
+export { default as SignUpButton }from "./SignUpButton";
+export { default as SignUpModel } from "./SignUpModel";
