+++ server/app/app_store/__init__.py
@@ -1,0 +1,1 @@
+from .functions import *
