+++ server/libs/model/login_request_model.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+class LoginRequestModel(BaseModel):
+    username_or_email: str
+    password: str
