+++ server/libs/model/login_response_model.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+from .user_model import UserModel
+
+class LoginResponseModel(BaseModel):
+    message: str
+    user: UserModel
+
