+++ server/libs/model/user_model.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel, EmailStr
+
+class UserModel(BaseModel):
+    username: str
+    email: EmailStr
+    password: str
