+++ server/libs/query/__init__.py
@@ -1,0 +1,1 @@
+from .generate_quiz_query import GenerateQuizQuery
