+++ server/libs/query/generate_quiz_query.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+class GenerateQuizQuery(BaseModel):
+    question_type: str
+    num_question: int
